Refactored actions to own file
